audio: a limiter instead of a clip, and the SB16 applies its mixer volumes
Carmageddon still crackled in races after both pacing fixes, and only
with its CD music playing: it was clipping. QEMU's mixer adds every
voice at full scale -- its sb16 stored the CT1745's volume registers and
applied none, the CD drive plays at mode page 0x0E's full 0xff -- and
its s16 conversion saturates the sum, so effects over CD music were
squared off on every peak. Windows' Volume Control sliders reached
nothing.

The player: the embed audiodev runs out.format=f32 (QEMU's float
conversion does not saturate) and the consumer has a 2 ms look-ahead
limiter, the identity below full scale; the ring is 128 KiB so it holds
as many milliseconds as before. A report line says when the guest's mix
went past full scale and how far it was turned down.

Patch 61: sb16 applies master x voice to its own voice (reapplied after
every AUD_open_out, which resets it) and mirrors the SB Pro registers
onto the SB16 ones both ways; the OPL3 and the CD drive's audio are
other devices' voices, reached through a small mixer-input registry in
the audio core (audio_mixin_*), the CD's through the 0x3C output
switches too. Reset values are 0 dB with the CD on rather than a cold
CT1745's -14 dB and muted CD, so a DOS game that never programs the
mixer sounds as it did; Windows' driver writes its own at boot.

Tests: the sb-mixer check (FM, master and SB Pro FM at -12 dB, each
-12.0 dB in QEMU's own wav); tools/audio-glitch-test.py gains cd and
cd+sb16 (CD-DA off a generated tone disc by PIO), CDAMP= (a loud
cd+sb16: 200 000 clipped samples per 5 s before, held at -2.4 dB with a
clean tone now), CDVOL= (the card's CD volume: -12.0 dB for 0xc8), and
fails on any clipped sample, which its click detector cannot see.
scripts/test.sh all: 57 passed.

+++ player/src/qemu_vm.rs
@@ -671,11 +671,14 @@ pub fn start(
         // pull; the consumer waits for the same amount before it plays. A
         // 5 ms mixer tick: a main loop held up by a 3D swap is paid back a
         // tick at a time, and finer ticks fit between the stalls
-        // (embed/embedaudio.c)
+        // (embed/embedaudio.c). f32, because QEMU's s16 output saturates
+        // a mix of voices that sums past full scale and its float output
+        // does not: the player's limiter turns it down instead
+        // (player/src/audio.rs)
         let cushion = crate::audio::cushion_ms() * 1000;
         args.push(format!(
             "embed,id=embed0,timer-period=5000,out.frequency={rate},out.channels=2,\
-             out.format=s16,out.buffer-length={cushion}"
+             out.format=f32,out.buffer-length={cushion}"
         ));
         // ring lives for the process; leak a strong ref for the C side
         let r = Arc::into_raw(ring);
